Move received CID stuffs to separate file
Final goal here is to have separate files for received_cid,
issued_cid, and retire_cid respectively. (The latter two will be
added later.)
