+++ lib/received_cid.c
@@ -21,7 +21,7 @@
  */
 
 #include <string.h>
-#include "quicly/cid.h"
+#include "quicly/received_cid.h"
 
 void quicly_received_cid_init(struct st_quicly_received_cid_set_t *set)
 {
