+++ t/test.h
@@ -56,6 +56,6 @@ void test_sentmap(void);
 void test_simple(void);
 void test_loss(void);
 void test_stream_concurrency(void);
-void test_cid(void);
+void test_received_cid(void);
 
 #endif
